Refactor satpy.writers for better import management
