+++ satpy/aux_download.py
@@ -20,8 +20,6 @@
 import logging
 import os
 
-import pooch
-
 import satpy
 
 logger = logging.getLogger(__name__)
@@ -105,6 +103,8 @@ def retrieve(cache_key, pooch_kwargs=None):
 
 
     """
+    import pooch
+
     pooch_kwargs = pooch_kwargs or {}
 
     path = satpy.config.get("data_dir")
@@ -121,6 +121,8 @@ def retrieve(cache_key, pooch_kwargs=None):
 
 
 def _retrieve_all_with_pooch(pooch_kwargs):
+    import pooch
+
     if pooch_kwargs is None:
         pooch_kwargs = {}
     path = satpy.config.get("data_dir")
@@ -225,7 +227,8 @@ def _find_registerable_files_readers(readers=None):
 
 def _find_registerable_files_writers(writers=None):
     """Load all writers so that files are registered."""
-    from satpy.writers import configs_for_writer, load_writer_configs
+    from satpy.writers.utils import configs_for_writer, load_writer_configs
+
     for writer_configs in configs_for_writer(writer=writers):
         try:
             load_writer_configs(writer_configs)
