+++ satpy/multiscene/_multiscene.py
@@ -31,7 +31,6 @@ import xarray as xr
 
 from satpy.dataset import DataID
 from satpy.scene import Scene
-from satpy.writers import get_enhanced_image, split_results
 
 try:
     import imageio
@@ -379,6 +378,8 @@ class MultiScene(object):
 
     def _distribute_save_datasets(self, scenes_iter, client, batch_size=1, **kwargs):
         """Distribute save_datasets across a cluster."""
+        from satpy.writers.utils import split_results
+
         def load_data(q):
             idx = 0
             while True:
@@ -460,6 +461,8 @@ class MultiScene(object):
 
     def _get_animation_info(self, all_datasets, filename, fill_value=None):
         """Determine filename and shape of animation to be created."""
+        from satpy.writers.utils import get_enhanced_image
+
         valid_datasets = [ds for ds in all_datasets if ds is not None]
         first_dataset = valid_datasets[0]
         last_dataset = valid_datasets[-1]
@@ -498,6 +501,8 @@ class MultiScene(object):
 
         Yet a single image frame from a dataset.
         """
+        from satpy.writers.utils import get_enhanced_image
+
         enh_args = enh_args.copy()  # don't change caller's dict!
         if "decorate" in enh_args:
             enh_args["decorate"] = self._format_decoration(
@@ -684,7 +689,7 @@ class MultiScene(object):
                 is a dask ``Client`` object then it will be used for
                 distributed computation.
             enh_args (Mapping): Optional, arguments passed to
-                :func:`satpy.writers.get_enhanced_image`.  If this includes a
+                :func:`satpy.writers.utils.get_enhanced_image`.  If this includes a
                 keyword "decorate", in any text added
                 to the image, string formatting will be applied based on
                 dataset attributes.  For example, passing
