+++ satpy/tests/features/steps/steps-save.py
@@ -21,6 +21,8 @@ from unittest.mock import patch
 
 from behave import given, then, use_step_matcher, when
 
+import satpy.writers.utils
+
 use_step_matcher("re")
 
 
@@ -49,7 +51,7 @@ def step_impl_scene_show(context):
 
     """
     with patch("trollimage.xrimage.XRImage.show") as mock_show:
-        context.scene.show("MyDataset")
+        satpy.writers.utils.show("MyDataset")
         mock_show.assert_called_once_with()
