+++ satpy/tests/modifier_tests/test_parallax.py
@@ -32,7 +32,7 @@ from pyresample import create_area_def
 
 import satpy.resample
 from satpy.tests.utils import xfail_skyfield_unstable_numpy2
-from satpy.writers import get_enhanced_image
+from satpy.writers.utils import get_enhanced_image
 
 # NOTE:
 # The following fixtures are not defined in this file, but are used and injected by Pytest:
