+++ satpy/tests/test_config.py
@@ -222,7 +222,7 @@ def _write_fake_writer_yaml(yaml_filename: str) -> None:
         comps_file.write(f"""
 writer:
     name: {writer_name}
-    writer: !!python/name:satpy.writers.Writer
+    writer: !!python/name:satpy.writers.base.Writer
 """)
 
 
@@ -336,13 +336,13 @@ class TestPluginsConfigs:
     @pytest.mark.parametrize("specified_writer", [None, "fake_writer"])
     def test_plugin_writer_configs(self, fake_writer_plugin_etc_path, specified_writer):
         """Test that writers can be loaded from plugin entry points."""
-        from satpy.writers import configs_for_writer
+        from satpy.writers.utils import configs_for_writer
         writer_yaml_path = fake_writer_plugin_etc_path / "writers" / "fake_writer.yaml"
         self._get_and_check_reader_writer_configs(specified_writer, configs_for_writer, writer_yaml_path)
 
     def test_plugin_writer_available_writers(self, fake_writer_plugin_etc_path):
         """Test that readers can be loaded from plugin entry points."""
-        from satpy.writers import available_writers
+        from satpy.writers.utils import available_writers
         self._check_available_component(available_writers, "fake_writer")
 
     @staticmethod
