+++ satpy/tests/writer_tests/test_cf.py
@@ -66,8 +66,8 @@ class TestCFWriter:
 
     def test_init(self):
         """Test initializing the CFWriter class."""
-        from satpy.writers import configs_for_writer
         from satpy.writers.cf_writer import CFWriter
+        from satpy.writers.utils import configs_for_writer
 
         CFWriter(config_files=list(configs_for_writer("cf"))[0])
