+++ satpy/tests/writer_tests/test_simple_image.py
@@ -66,8 +66,8 @@ class TestPillowWriter(unittest.TestCase):
         """Test writing datasets with delayed computation."""
         from dask.delayed import Delayed
 
-        from satpy.writers import compute_writer_results
         from satpy.writers.simple_image import PillowWriter
+        from satpy.writers.utils import compute_writer_results
         datasets = self._get_test_datasets()
         w = PillowWriter(base_dir=self.base_dir)
         res = w.save_datasets(datasets, compute=False)
