+++ satpy/writers/mitiff.py
@@ -25,7 +25,8 @@ import numpy as np
 from PIL import Image, ImagePalette
 
 from satpy.dataset import DataID, DataQuery
-from satpy.writers import ImageWriter, get_enhanced_image
+from satpy.writers.base_image import ImageWriter
+from satpy.writers.utils import get_enhanced_image
 
 IMAGEDESCRIPTION = 270
